Make it work after /upgrade
After running /upgrade, the script is reloaded, but the old buffer still
exists so a new one can't be created. This meant himan_buffer was set to
None which made the lines go to the core buffer instead. To fix this,
search for an existing buffer on startup and use that if it exists.